perf: 优化配置文件示例和调整无用代码
